feat: omit nullish serialized values
Doing this simplifies some of the implementation. Composed
serializers no longer throw if they do not return a value. It also
removes RichTextError from the library.

This allows users to omit element types from a serializer if they are
not needed. For example, if a user only wants to render `heading1`
elements, the user can provide a serializer that only returns values for
`heading1` nodes. All other types can be ignored and will not be
included in the output.